requirements.txt
torch
transformers
huggingface_hub
accelerate
python-dotenv
